+++ src/components/studyGroup/TagBadge.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { Badge } from "./TagBadge.styled";
+import { Badge } from "./styles/TagBadge.styled";
 
 const TagBadge = ({ tag }) => <Badge>#{tag}</Badge>;
