+++ src/components/studyGroup/styles/StudyGroupCard.styled.js
