+++ src/components/studyGroup/styles/TagBadge.styled.js
