+++ src/pages/StudyGroup/styles/StudyGroupList.styled.js
